Kommentare!

+++ Eclipse/src/PlayerShip.java
@@ -1,23 +1,29 @@
 import java.awt.Graphics;
 import java.awt.event.KeyEvent;
 
-
+/* Stellt das Schiff eines menschlichen Spielers dar
+ * 
+ */
 public abstract class PlayerShip {
 	
-	protected int x;
-	protected int y;
+	protected float x;
+	protected float y;
 	protected int width;
 	protected int height;
 	
-	public PlayerShip(int x, int y, int width, int height) {
+	//Konstruktor. Selbsterklärend!?
+	//X und Y sind als Floats, um Zwischenwerte zu erhalten, die für Berechnungen wichtig sind, jedoch nicht angezeigt werden
+	public PlayerShip(float x, float y, int width, int height) {
 		this.x = x;
 		this.y = y;
 		this.width = width;
 		this.height = height;
 	}
 	
+	//Zeichen-Methode.
 	public abstract void paintShip(Graphics g);
 	
+	//Updated die Position des Schiffes
 	public abstract void update();
 	
 	public abstract void keyPressed(KeyEvent event);
